feat(api): create draft-modules route
