+++ api/lib/application/modules/draft-modules.js
@@ -1,0 +1,39 @@
+import Joi from 'joi';
+
+import { draftModuleSerializer } from '../../infrastructure/serializers/jsonapi/index.js';
+import { createDraftModule } from '../../domain/usecases/index.js';
+
+export function register(server) {
+  server.route([
+    {
+      method: 'POST',
+      path: '/api/draft-modules',
+      config: {
+        validate: {
+          payload: Joi.object({
+            data: Joi.object({
+              type: Joi.string().valid('draft-modules').required(),
+              attributes: Joi.object({
+                'internal-title': Joi.string().required(),
+                title: Joi.string().required(),
+                'is-beta': Joi.boolean().required(),
+                slug: Joi.string().required(),
+                visibility: Joi.string().required(),
+                details: Joi.object().required(),
+                sections: Joi.array().required(),
+                glossary: Joi.array().required(),
+              }).required(),
+            }).required(),
+          }).required(),
+        },
+        handler: async (request, h) => {
+          const module = await draftModuleSerializer.deserialize(request.payload);
+          const savedModule = await createDraftModule(module);
+          return h.response(draftModuleSerializer.serialize(savedModule)).code(201);
+        },
+      },
+    },
+  ]);
+}
+
+export const name = 'draft-modules';
